Parse down to the City/Roadway level
Now able to parse down to the City/Roadway level. For example,
I tested using Cambridge + Memorial Drive. There are some quirks
to figure out, like how we need to capitalize the City name, and
how I need to add a space in front of the roadway currently. I'm
guessing I need to tweak the filtering to be more "grep like"
instead of using ==. I'll need to research how to do that with
pandas.

I also (in a future commit) want to generate stats for the City and
Roadway. Stats like:
* number of pedestrian crashes found for City, plus City/Roadway
* number of cycling crashes found
* number of pedestrian + cycling crashes found

I also want to make it possible to generate filtered CSV files
for every town in Mass, and then every Town/Roadway combination
too. This will take some thinking, but perhaps I can get a list
of unique Cities, and Unique roadways, then just do a double for
loop over each. I could parallelize this to make it multi-threaded
using the multiprocessing library. Something to chew on.