Refactor to use functions + Reuse existing merged CSV file
This commit refactors some early code to use a "mergeCsvFiles"
function, so that we can easily bypass this logic if we decide to
reuse the existing merged CSV file. This should speed up execution
of the script greatly as we'll no longer need to wait for ~2GB
of CSV files to be combined; instead we can just read it straight
from disk.

I should also probably make it so we can reuse filtered City CSV
files too so if someone passes in "Somerville", if the
"massDOT_data_SOMERVILLE.csv" file already exists, we could just
filter the Roadway on that. Likewise, if both files already exist
we could warn the user.