Update README with notes
Added more notes to the README on how to get MassDOT IMPACT data off their
crash portal. Will eventually get my copy of the data onto somewhere public
that isn't Google Drive or Dropbox so I can share it with a wider audience.

That'll be tricky since this data is about 2GB worth for 2010 - 2021.